Renamed "basic" to "classic" in plan scheme types
